+++ R/wiki_onthisday.R
@@ -1,17 +1,19 @@
-#' Generate random 'on this day' fact from current or historic Wikipedia main page
+#' Generate random 'on this day' facts from current or historic Wikipedia main page
 #'
 #' @description
-#' `wiki_onthisday()` generates a random 'on this day' fact from a current or historic Wikipedia main page
+#' `wiki_onthisday()` generates random 'on this day' facts from a current or historic Wikipedia main page
 #'
-#' @param date A date string of the form YYYY-MM-DD.  Default value is today's date.
-#' @return A string with a random 'on this day' fact from Wikipedia's main page if it exists for the date specified - otherwise "I got nothin'"
+#' @param n_facts An integer determining the number of facts that will be generated, up to a limit of the maximum facts for the date specified.
+#' @param date A date string of the form YYYY-MM-DD.  Default value is yesterday's date.
+#' @param bare_fact Logical.  Determining whether the fact should be quoted as is or surrounded by a preamble and courtesy statement.
+#'
+#' @return A vector of strings with random 'on this day' facts from Wikipedia's main page if it exists for the date specified - otherwise "I got nothin'"
 #'
 #' @examples
 #' wiki_onthisday(date = '2020-05-02')
 
-wiki_onthisday <- function(date = Sys.Date()) {
+wiki_onthisday <- function(n_facts = 1L, date = Sys.Date() - 1, bare_fact = FALSE) {
 
-  # get url from input and read html
   # get url from input and read html
   date <- as.Date(date)
   date1 <- format(date, "%Y_%B_")
@@ -31,11 +33,13 @@ wiki_onthisday <- function(date = Sys.Date()) {
       rvest::html_nodes("li") %>%
       rvest::html_text()
 
-    otd <- otd[grepl("^\\d{4}", otd)] %>%
-      sample(1)
+    n <- min(n_facts, length(otd))
 
-    if (date == Sys.Date()) {
-      paste("Did you know that on this day in", otd, "(Courtesy of Wikipedia)")
+    otd <- otd[grepl("^\\d{4}", otd)] %>%
+      sample(n)
+
+    if(bare_fact == TRUE) {
+      otd
     } else {
       paste0("Did you know that on ", format(date, "%B"), " ", date2, " in ", otd, " (Courtesy of Wikipedia)")
     }