+++ tests/testthat/test-wiki_search.R
@@ -1,0 +1,4 @@
+test_that("wiki_search() works", {
+  wiki_search('R Programming Language') %>%
+    testthat::expect_silent()
+})