+++ NEWS.md
@@ -1,5 +1,10 @@
 # wikifacts 0.2.9000
 
+* Generate multiple facts within functions
+* Move default date to yesterday to avoid timezone issues
+* Add ability to display bare facts without cosmetic wrapping
+* Add ability for `wiki_randomfact()` to select the type of fact to randomize
+
 # wikifacts 0.2
 
 * Allows facts from historic Wikipedia main pages by adding a `date` argument to functions.