+++ NEWS.md
@@ -4,6 +4,7 @@
 * Move default date to yesterday to avoid timezone issues
 * Add ability to display bare facts without cosmetic wrapping
 * Add ability for `wiki_randomfact()` to select the type of fact to randomize
+* Remove date argument from `wiki_randomfact()`
 
 # wikifacts 0.2