+++ NEWS.md
@@ -1,11 +1,11 @@
-# wikifacts 0.2.9000
+# wikifacts 0.3.0
 
 * Generate multiple facts within functions (#5)
 * Move default date to yesterday to avoid timezone issues
 * Add ability to display bare facts without cosmetic wrapping
 * Add ability for `wiki_randomfact()` to select the type of fact to randomize
 * Remove date argument from `wiki_randomfact()`
-* Add `wiki_search()`
+* Add `wiki_search()` to open browser with results of search
 
 # wikifacts 0.2